pdf2latex_pipeline/pdf2latex/requirements.txt
# Core deep learning
torch>=2.2.0
torchvision>=0.17.0

# Transformers & tokenisation
transformers>=4.40.0
tokenizers>=0.19.0

# PDF rendering
pymupdf>=1.23.0       # provides `fitz`

# Metrics
editdistance>=0.8.0
nltk>=3.8.0

# Config & utilities
pyyaml>=6.0
tqdm>=4.66.0

# Optional: experiment tracking
wandb>=0.17.0

# Optional: distributed training
accelerate>=0.30.0

# Optional: faster compilation checks
# tectonic               # alternative to pdflatex (statically linked)
